requirement.txt
chainlit==1.0.0 
google-generativeai>=0.5.0 
sentence-transformers==2.6.1
faiss-cpu==1.12.0
pandas==2.0.3 
python-dotenv==1.0.0 